+++ src/main/java/com/telepathicgrunt/world_blender/WorldBlender.java
@@ -29,7 +29,6 @@ public class WorldBlender implements ModInitializer {
 		WB_CONFIG = AutoConfig.getConfigHolder(WBConfig.class).getConfig();
 
 		WBBlocks.register();
-		WBPortalSpawning.generateRequiredBlockList(WB_CONFIG.WBPortalConfig.requiredBlocksInChests);
 
 		WorldBlender.reserveBiomeIDs();
 		WBFeatures.registerFeatures();